+++ pkg/unittest/test_runner_test.go
@@ -2,6 +2,7 @@ package unittest_test
 
 import (
 	"bytes"
+	"fmt"
 	"os"
 	"path/filepath"
 	"regexp"
@@ -12,10 +13,10 @@ import (
 	"testing/fstest"
 
 	"github.com/bradleyjkemp/cupaloy/v2"
-	"github.com/helm-unittest/helm-unittest/internal/printer"
+	"github.com/helm-unittest/helm-unittest/internal/common"
 	. "github.com/helm-unittest/helm-unittest/pkg/unittest"
+	"github.com/helm-unittest/helm-unittest/pkg/unittest/printer"
 	"github.com/stretchr/testify/assert"
-	"sigs.k8s.io/yaml"
 )
 
 var sectionBeginPattern = regexp.MustCompile("( PASS | FAIL |\n*###|\n*Charts:|\n*Snapshot Summary:)")
@@ -277,6 +278,55 @@ func TestV3RunnerOkWithSchemaValidation(t *testing.T) {
 	cupaloy.SnapshotT(t, makeOutputSnapshotable(buffer.String())...)
 }
 
+func TestV3RunnerOk_With_FailFast_NoPanic(t *testing.T) {
+	buffer := new(bytes.Buffer)
+	runner := TestRunner{
+		Printer:   printer.NewPrinter(buffer, nil),
+		TestFiles: []string{testTestFiles},
+	}
+	cases := []struct {
+		chartPath []string
+		failFast  bool
+	}{
+		{
+			chartPath: []string{testV3WithFailingTemplateChart},
+			failFast:  true,
+		},
+		{
+			chartPath: []string{testV3WithFailingTemplateChart},
+			failFast:  false,
+		},
+		{
+			chartPath: []string{testV3InvalidBasicChart},
+			failFast:  true,
+		},
+		{
+			chartPath: []string{testV3InvalidBasicChart},
+			failFast:  false,
+		},
+	}
+	for _, tt := range cases {
+		t.Run(fmt.Sprintf("chart %s fail fast: %v", tt.chartPath[0], tt.failFast), func(t *testing.T) {
+			runner.Failfast = tt.failFast
+			result := runner.RunV3([]string{testV3WithFailingTemplateChart})
+			assert.True(t, result)
+		})
+	}
+}
+
+func TestV3RunnerOkWithDocumentSelect(t *testing.T) {
+	buffer := new(bytes.Buffer)
+	runner := TestRunner{
+		Printer:   printer.NewPrinter(buffer, nil),
+		TestFiles: []string{testTestFiles},
+	}
+	passed := runner.RunV3([]string{testV3WithDocumentSelectorChart})
+	assert.True(t, passed, buffer.String())
+	fmt.Println(buffer.String())
+	assert.Contains(t, buffer.String(), "Test Suites: 7 passed, 7 total")
+	assert.Contains(t, buffer.String(), "Tests:       10 passed, 10 total")
+}
+
 func TestV3RunnerOkWithTestSkipped(t *testing.T) {
 	suiteDoc := `
 suite: test suite with subchart
@@ -291,8 +341,7 @@ tests:
       - matchSnapshot: {}
 `
 	testSuite := TestSuite{}
-	err := yaml.Unmarshal([]byte(suiteDoc), &testSuite)
-	assert.Nil(t, err)
+	common.YmlUnmarshalTestHelper(suiteDoc, &testSuite, t)
 
 	buffer := new(bytes.Buffer)
 	runner := TestRunner{