+++ pkg/unittest/utils_test.go
@@ -2,13 +2,14 @@ package unittest_test
 
 import (
 	"bytes"
+	"fmt"
 	"os"
 	"path/filepath"
 	"testing"
 
 	"github.com/helm-unittest/helm-unittest/internal/common"
-	"github.com/helm-unittest/helm-unittest/internal/printer"
 	. "github.com/helm-unittest/helm-unittest/pkg/unittest"
+	"github.com/helm-unittest/helm-unittest/pkg/unittest/printer"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -67,4 +68,141 @@ func TestV3RunnerWith_Fixture_Chart_ErrorWhenMetaCharacters(t *testing.T) {
 	}
 	passed := runner.RunV3([]string{"testdata/chart01"})
 	assert.True(t, passed, buffer.String())
+}
+
+func TestV3RunnerWith_Fixture_Chart_FailFast(t *testing.T) {
+	cases := []struct {
+		chart      string
+		failFast   bool
+		testFlavor string
+		expected   []string
+	}{
+		{
+			chart:      "testdata/chart-fail-fast",
+			failFast:   true,
+			testFlavor: "case1",
+			expected: []string{
+				"FAIL  a fail-fast first test",
+				"Test Suites: 1 failed, 0 passed, 1 total",
+				"Tests:       1 failed, 1 passed, 2 total",
+			},
+		},
+		{
+			chart:      "testdata/chart-fail-fast",
+			failFast:   false,
+			testFlavor: "case1",
+			expected: []string{
+				"FAIL  a fail-fast first test",
+				"PASS  b fail-fast second test",
+				"Test Suites: 1 failed, 1 passed, 2 total",
+				"Tests:       1 failed, 4 passed, 5 total",
+			},
+		},
+		{
+			chart:      "testdata/chart-fail-fast",
+			failFast:   false,
+			testFlavor: "case2",
+			expected: []string{
+				"PASS  a fail-fast first test all pass",
+				"FAIL  b fail-fast second test",
+				"Test Suites: 1 failed, 1 passed, 2 total",
+				"Tests:       1 failed, 5 passed, 6 total",
+			},
+		},
+		{
+			chart:      "testdata/chart-fail-fast",
+			failFast:   true,
+			testFlavor: "case2",
+			expected: []string{
+				"PASS  a fail-fast first test all pass",
+				"FAIL  b fail-fast second test",
+				"Test Suites: 1 failed, 1 passed, 2 total",
+				"Tests:       1 failed, 5 passed, 6 total",
+			},
+		},
+	}
+	for _, tt := range cases {
+		t.Run(fmt.Sprintf("chart %s with %s fail fast %v", tt.chart, tt.testFlavor, tt.failFast), func(t *testing.T) {
+			buffer := new(bytes.Buffer)
+			runner := TestRunner{
+				Printer:   printer.NewPrinter(buffer, nil),
+				TestFiles: []string{fmt.Sprintf("tests/*-%s_test.yaml", tt.testFlavor)},
+				Failfast:  tt.failFast,
+			}
+			_ = runner.RunV3([]string{"testdata/chart-fail-fast"})
+			for _, e := range tt.expected {
+				assert.Contains(t, buffer.String(), e)
+			}
+		})
+	}
+}
+
+func TestV3RunnerWith_Fixture_Chart_YamlSeparator(t *testing.T) {
+	buffer := new(bytes.Buffer)
+	runner := TestRunner{
+		Printer:   printer.NewPrinter(buffer, nil),
+		TestFiles: []string{"tests/*_test.yaml"},
+		Strict:    false,
+	}
+	_ = runner.RunV3([]string{"testdata/chart-yaml-separator"})
+	assert.Contains(t, buffer.String(), "Test Suites: 5 passed, 5 total")
+	assert.Contains(t, buffer.String(), "Tests:       6 passed, 6 total")
+}
+
+func TestV3RunnerWith_Fixture_Chart_DocumentSelector(t *testing.T) {
+	cases := []struct {
+		chart      string
+		testFlavor string
+		expected   []string
+	}{
+		{
+			chart:      "testdata/chart-document-selector",
+			testFlavor: "case1-error",
+			expected: []string{
+				"### Error:  empty 'documentSelector.path' not supported",
+			},
+		},
+		{
+			chart:      "testdata/chart-document-selector",
+			testFlavor: "case2-error",
+			expected: []string{
+				"### Error:  empty 'documentSelector.value' not supported",
+			},
+		},
+		{
+			chart:      "testdata/chart-document-selector",
+			testFlavor: "case3-error",
+			expected: []string{
+				"Template:\tdocument-selector/templates/cfg01.yaml",
+				"Path:\tkind expected to exist",
+			},
+		},
+		{
+			chart:      "testdata/chart-document-selector",
+			testFlavor: "case4-error",
+			expected: []string{
+				"Path:\tkind expected to exists",
+			},
+		},
+		{
+			chart:      "testdata/chart-document-selector",
+			testFlavor: "case1-ok",
+			expected: []string{
+				"Test Suites: 2 passed, 2 total",
+			},
+		},
+	}
+	for _, tt := range cases {
+		t.Run(fmt.Sprintf("chart %s with %s", tt.chart, tt.testFlavor), func(t *testing.T) {
+			buffer := new(bytes.Buffer)
+			runner := TestRunner{
+				Printer:   printer.NewPrinter(buffer, nil),
+				TestFiles: []string{fmt.Sprintf("tests/%s_test.yaml", tt.testFlavor)},
+			}
+			_ = runner.RunV3([]string{tt.chart})
+			for _, e := range tt.expected {
+				assert.Contains(t, buffer.String(), e)
+			}
+		})
+	}
 }